Lots of automation work Config fix
git-svn-id: http://svn.lynbrookrobotics.com/code15@120 9f8849d2-cf13-4b18-a094-84ffabbe6057

+++ Robot/LRT15/src/com/lynbrookrobotics/frc2015/automation/ReleaseStack.java
@@ -9,7 +9,7 @@ public class ReleaseStack extends Automation implements Configurable {
 
 	private ElevatorData elevatorData;
 	private float startingPosition;
-	private float dropHeight;
+	private double dropHeight;
 
 	public ReleaseStack() {
 		super("ReleaseStack");
@@ -40,7 +40,7 @@ public class ReleaseStack extends Automation implements Configurable {
 	@Override
 	protected boolean Run() {
 		elevatorData.setControlMode(ControlMode.POSITION);
-		elevatorData.setPosition(startingPosition - dropHeight);
+		elevatorData.setPosition((float) (startingPosition - dropHeight));
 		return true;
 	}
